+++ CHANGELOG.md
@@ -2,13 +2,15 @@
 
 Version 2.4.1:
 
+* Deprecated `grizzled.config.Configuration`. If you need a good configuration
+  file format, use YAML or [HOCON](https://github.com/typesafehub/config).
 * Various fixes for Windows, including:
     - Changed various classes to use use the `line.separator`
       property, rather than a hard-coded "\n", when adding line terminators.
     - Changed how `grizzled.net.URLUtil` resolves parent directories.
-* Updated to ScalaTest 3.0.0-RC1.
 * Removed `BrainDeadHTTPServer` (used during testing) as it overly complicates
   testing. `file:` URLs are now used instead.
+
 
 Version 2.4.0: